refactor(hopr-node): retire hopr-reference, introduce impls/node husk
Rename hopr/reference → impls/node (package hopr-node). Strip the crate
to the minimum assembly required by integration tests and remove all
production-glue code that is now handled by the hopr-lib builder.

Changes:
- Delete configure_node, create_default_blokli_connector, build_edge,
  build_full_with_session_server (Level-1 wrappers): unused by any
  consumer; the peer-discovery bridge now lives in the hopr-lib builder
  (PR #1).
- Keep build_edge_with_chain + build_full_with_chain (Level-2 builders)
  and their FullHopr / EdgeHopr / SharedTicketManager type aliases:
  edge-client and hoprd depend on these until their own migration PRs.
- Replace inline ticket-seeding loops with ticket_factory_from_chain /
  ticket_manager_from_chain helpers from impls/ticket-manager (PR #1).
- Wire with_peer_discovery_tx(peer_discovery_tx) in both builders so
  chain-event announcements reach libp2p (PR #1).
- Move session_integration_tests.rs + telemetry_integration.rs (and
  their insta snapshots) into hopr/hopr-lib/tests/: they exercise only
  hopr_lib::exports and the metrics provider, not the node assembly.
- Add the required dev-deps (opentelemetry_sdk, rstest, hopr-api with
  types-random/telemetry) to hopr-lib/Cargo.toml.
- Update workspace Cargo.toml: members hopr/reference → impls/node,
  dep alias hopr-reference → hopr-node.
