+++ Cargo.toml
@@ -18,7 +18,7 @@ members = [
   "transport/tag-allocator",
   "utils-session",
   # implementations
-  "hopr/reference",
+  "impls/node",
   "impls/connector",
   "impls/session-server-forwarder",
   "impls/ct/immediate",
@@ -180,7 +180,7 @@ hopr-transport-tag-allocator = { path = "transport/tag-allocator", default-featu
 hopr-chain-connector = { path = "impls/connector" }
 
 # referential implementations
-hopr-reference = { path = "hopr/reference", default-features = false }
+hopr-node = { path = "impls/node", default-features = false }
 hopr-session-server-forwarder = { path = "impls/session-server-forwarder", default-features = false }
 hopr-transport-p2p = { path = "impls/transport/p2p", default-features = false }
 hopr-ct-immediate = { path = "impls/ct/immediate", default-features = false }
