+++ hopr/hopr-lib/tests/telemetry_integration.rs
