--- hopr/reference/.gitignore
@@ -1,6 +1,0 @@
-/target
-**/*.rs.bk
-Cargo.lock
-bin/
-pkg/
-wasm-pack.log
