+++ impls/node/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "hopr-reference"
-version = "4.2.3"
+name = "hopr-node"
+version = "0.1.0"
 authors = ["HOPR Association <tech@hoprnet.org>"]
 edition.workspace = true
 rust-version.workspace = true
-description = "Referentail omplementation of the hopr-lib object using canonical trait implementations"
+description = "Canonical assembly of hopr-lib with concrete impls; integration-test rig for the HOPR protocol"
 homepage = "https://hoprnet.org/"
 repository = "https://github.com/hoprnet/hoprnet"
 license = "GPL-3.0-only"
@@ -73,7 +73,7 @@ hopr-session-server-forwarder = { workspace = true, optional = true }
 hopr-chain-connector = { workspace = true }
 hopr-ct-full-network = { workspace = true }
 hopr-network-graph = { workspace = true, features = ["petgraph", "graph-api"] }
-hopr-ticket-manager = { workspace = true }
+hopr-ticket-manager = { workspace = true, features = ["redb"] }
 hopr-transport-p2p = { workspace = true }
 
 [dev-dependencies]
@@ -83,7 +83,7 @@ serial_test = { workspace = true }
 test-log = { workspace = true }
 tokio = { workspace = true }
 
-hopr-reference = { workspace = true, features = ["testing"] }
+hopr-node = { workspace = true, features = ["testing"] }
 hopr-api = { workspace = true, features = ["types-random", "types-telemetry"] }
 opentelemetry_sdk = { workspace = true, features = [
   "metrics",
