+++ impls/node/tests/api_general.rs
@@ -1,6 +1,6 @@
 use anyhow::Context;
 use hopr_lib::api::{PeerId, chain::ChainKeyOperations, node::HasChainApi};
-use hopr_reference::testing::fixtures::{ClusterGuard, TEST_GLOBAL_TIMEOUT, size_2_cluster_fixture as cluster};
+use hopr_node::testing::fixtures::{ClusterGuard, TEST_GLOBAL_TIMEOUT, size_2_cluster_fixture as cluster};
 use rstest::*;
 use serial_test::serial;
